Replace setImmediate with process.nextTick
setImmediate runs after almost a whole event loop cycle (except close callbacks)
which means that what we put in setImmediate is affected by blocked "event loop
latency"; i.e. if anything between the point setImmediate is executed and the "check"
section of event loop blocks the event loop (for example one of the pending callbacks
in poll section), then latency of limiter will increase; even for the case of non-ratelimiting
(unlimited: true). This PR replaces setImmediate by process.nextTick, nextTick is executed
after current function but before the event loop is allowed to continue, providing a very low
latency for the case of not-rate-limit.